API changes

+++ app/database/policy_database.py
@@ -10,28 +10,29 @@ class PolicyDatabase:
         self.database = TinyDB(database_url)
         self.store = Query()
 
-    def get_policy(self, policy_name: str) -> dict:
-        return self.database.get(self.store.name == policy_name)
+    def get_policy(self, policy_name: str, owner: str) -> dict:
+        return self.database.get(self.store.name == policy_name and self.store.owner == owner)
 
-    def add_policy(self, policy: RequestObject) -> dict:
-        if self.get_policy(policy["name"]):
+    def add_policy(self, policy: dict, owner: str) -> dict:
+        if self.exists(policy["name"], owner):
             raise HTTPException(
                 status_code=409, detail="Rules with the same name already exist"
             )
         self.database.insert(policy)
         return policy
 
-    def update_policy(self, policy_name: str, policy: dict) -> None:
-        self.database.update(policy, self.store.name == policy_name)
+    def update_policy(self, policy_name: str, policy: dict, owner: dict) -> None:
+        self.database.update(policy, self.store.name == policy_name and self.store.owner == owner)
 
-    def exists(self, policy_name: str) -> bool:
-        return self.database.contains(self.store.name == policy_name)
+    def exists(self, policy_name: str, owner: str) -> bool:
+        doc = self.database.get(self.store.name == policy_name and self.store.owner == owner)
+        return doc["name"] == policy_name
 
-    def delete_policy(self, policy_name: str) -> None:
-        self.database.remove(self.store.name == policy_name)
+    def delete_policy(self, policy_name: str, owner: str) -> None:
+        self.database.remove(self.store.name == policy_name and self.store.owner == owner)
 
-    def get_policies(self) -> list:
-        return self.database.all()
+    def get_policies(self, owner: str) -> list:
+        return self.database.search(self.store.owner == owner)
 
 
 def get_db() -> PolicyDatabase:
